big lump of patch that moves xml schema functionality out of the wsdl package.
