+++ src/rpclib/interface/xml_schema/model/primitive.py
@@ -1,4 +1,22 @@
 
+#
+# rpclib - Copyright (C) Rpclib contributors.
+#
+# This library is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License as published by the Free Software Foundation; either
+# version 2.1 of the License, or (at your option) any later version.
+#
+# This library is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public
+# License along with this library; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+#
+
 import rpclib.const.xml_ns
 _ns_xs = rpclib.const.xml_ns.xsd
 
@@ -31,6 +49,8 @@ def string_get_restriction_tag(interface, cls):
         pattern = etree.SubElement(restriction, '{%s}pattern' % _ns_xs)
         pattern.set('value', cls.Attributes.pattern)
 
+    return restriction
+
 def decimal_get_restriction_tag(interface, cls):
     restriction = simple_get_restriction_tag(interface, cls)
 
@@ -50,3 +70,4 @@ def decimal_get_restriction_tag(interface, cls):
         min_l = etree.SubElement(restriction, '{%s}maxInclusive' % _ns_xs)
         min_l.set('value', str(cls.Attributes.le))
 
+    return restriction
